requirements.txt
Django==4.2.7
firebase-admin==6.2.0
python-decouple==3.8
requests==2.31.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
gunicorn==21.2.0
whitenoise==6.5.0